🎨 _profileContext 파일 삭제
- 이벤트는 _events.ts로
- MyProfileContext는 layout.tsx 로 이동
