+++ src/app/main/questions/page.tsx
@@ -4,11 +4,11 @@ import Question from '@/app/_components/question';
 import { useContext, useEffect, useRef, useState } from 'react';
 import DialogModalTwoButton from '@/app/_components/modalTwoButton';
 import DialogModalLoadingOneButton from '@/app/_components/modalLoadingOneButton';
-import { MyProfileContext } from '@/app/main/_profileContext';
 import { questionDto } from '@/app/_dto/questions/question.dto';
 import { MyQuestionEv } from '../_events';
 import { Logger } from '@/utils/logger/Logger';
 import { QuestionDeletedPayload } from '@/app/_dto/websocket-event/websocket-event.dto';
+import { MyProfileContext } from '@/app/main/layout';
 
 const fetchQuestions = async (): Promise<questionDto[] | null> => {
   const res = await fetch('/api/db/questions');
